+++ R/inflate.R
@@ -5,8 +5,10 @@
 #' @param rmd Path to Rmarkdown file to inflate
 #' @param check Logical. Whether to check package after Rmd inflating
 #' @param document Logical. Whether to document your package using \code{\link[attachment:att_amend_desc]{att_amend_desc}}
+#' @param overwrite Logical. Whether to overwrite vignette and functions if already exists.
 #'
 #' @importFrom parsermd parse_rmd as_tibble
+#' @importFrom utils getFromNamespace
 #' @return
 #' Package structure. Return path to current package.
 #' @export
@@ -31,7 +33,9 @@
 #' # usethis::use_git_ignore("docs")
 #' # Delete dummy package
 #' unlink(dummypackage, recursive = TRUE)
-inflate <- function(pkg = ".", rmd = file.path("dev", "dev_history.Rmd"), name = "exploration", check = TRUE, document = TRUE) {
+inflate <- function(pkg = ".", rmd = file.path("dev", "dev_history.Rmd"),
+                    name = "exploration", check = TRUE, document = TRUE,
+                    overwrite = c("ask", "yes", "no")) {
   old <- setwd(pkg)
   on.exit(setwd(old))
 
@@ -69,6 +73,24 @@ inflate <- function(pkg = ".", rmd = file.path("dev", "dev_history.Rmd"), name =
     stop(rmd, " does not exists, please use fusen::add_dev_history() to create it.")
   }
 
+  # Are you sure ?
+  overwrite <- match.arg(overwrite)
+  cleaned_name <- asciify_name(name)
+  vignette_path <- file.path(pkg, "vignettes", paste0(cleaned_name, ".Rmd"))
+  if (file.exists(vignette_path)) {
+    if (overwrite == "ask") {
+      rm_exist_vignette <- getFromNamespace("can_overwrite", "usethis")(vignette_path)
+    } else {
+      rm_exist_vignette <- ifelse(overwrite == "yes", TRUE, FALSE)
+    }
+    if (rm_exist_vignette) {
+      file.remove(vignette_path)
+    } else {
+      stop("Vignette already exists, anwser 'yes' to the previous question",
+           " or set inflate(..., overwrite = 'yes') to always overwrite.")
+    }
+  }
+
   # Create NAMESPACE
   namespace_file <- file.path(pkg, "NAMESPACE")
   if (!file.exists(namespace_file)) {
@@ -96,11 +118,13 @@ inflate <- function(pkg = ".", rmd = file.path("dev", "dev_history.Rmd"), name =
 
   # Check
   if (isTRUE(check)) {
-    rcmdcheck::rcmdcheck(pkg)
+    res <- rcmdcheck::rcmdcheck(pkg)
+    print(res)
   }
 
   pkg
 }
+
 
 #' Create function code, doc and tests ----
 #' @param parsed_tbl tibble of a parsed Rmd
@@ -162,7 +186,7 @@ get_functions <- function(parsed_tbl) {
       # find function name
       fun_name <- stringr::str_extract(
         code[grep("function(\\s*)\\(", code)],
-        "\\w*(?=(\\s*)(<-|=)(\\s*)function)"
+        "[\\w[.]]*(?=(\\s*)(<-|=)(\\s*)function)"
       ) %>%
         gsub(" ", "", .) # remove spaces
 
@@ -397,14 +421,28 @@ create_vignette <- function(parsed_tbl, pkg, name) {
   ]
 
   # Make chunk names unique
-  # vignette_tbl[["label"]] <- ifelse(
+  # vignette_tbl[["label"]][grepl("unnamed", vignette_tbl[["label"]])] <-
+  #   gsub("unnamed-", "parsermd-", vignette_tbl[["label"]][grepl("unnamed", vignette_tbl[["label"]])])
   #   is.na(vignette_tbl[["label"]]) & vignette_tbl[["type"]] == "rmd_chunk",
   #                                   gsub("[.]+", "-", make.names(name)),
   #                                   vignette_tbl[["label"]])
   #
   # vignette_tbl[["label"]] <- make.unique(vignette_tbl[["label"]], sep = "-")
-  # # Not re-used in as_document()
+  # # /!\ Not re-used in as_document(), this must be in ast
 
+  # ast <- vignette_tbl[["ast"]][[21]]
+
+  # To correct for {parsermd} unnamed attribution
+  fix_unnamed_chunks <- function(ast) {
+    if (inherits(ast, "rmd_chunk") && grepl("unnamed-chunk-", ast[["name"]])) {
+      ast[["name"]] <- gsub("unnamed-", "parsermd-", ast[["name"]])
+    }
+    ast
+  }
+
+  ast_class <- class(vignette_tbl[["ast"]])
+  vignette_tbl[["ast"]] <- lapply(vignette_tbl[["ast"]], fix_unnamed_chunks)
+  class(vignette_tbl[["ast"]]) <- ast_class
 
   cleaned_name <- asciify_name(name)